Added an optimized Dockerfile
The Dockerfile builds the final image in two stages.
The first stage pulls an Alpine Linux/Go-based image and builds
the server binaries in it.
The second stage pulls an Alpine Linux-based image that allows
decreasing the final image's size and copies the binaries built in the
first stage to it.