Cut out "generic" data handling where appropriate.
For the most part, the generic interface{} handling code has been
replaced with code that works with concrete types. This has the somewhat
unfortunate effect of bloating the code due to functions that are
essentially the same, but deal with slightly different data structures.
On the flip side, the code has also become much more readable.

Now to update the client application.